Add new search, delete, update, and DTO functionality to Movie Controller
This commit adds significant new functionality to the Movie Controller, including the ability to search for movies by director and year, genre and minimum duration, as well as the ability to count the total number of movies in the database. Additionally, it adds new delete methods for removing data by ID, list of IDs, movie and list of movies, as well as an update method for updating movie information. To support these changes, two new DTOs have been added: MovieListDTO for returning a list of movies with selected fields and CountDTO for returning the total count of movies in the database. These changes significantly enhance the flexibility of the Movie Controller and improve its ability to serve client requests.